IstanbulMedic-agent/requirements.txt
openai
openai-agents
langchain
langgraph
pinecone-client
python-dotenv
fastapi
twilio